# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Addon": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"authors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"options": {
						"properties": {
							"tags": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"optifine": {
								"type": "boolean"
							},
							"comments": {
								"type": "boolean"
							}
						},
						"required": [
							"tags",
							"optifine",
							"comments"
						],
						"type": "object"
					},
					"approval": {
						"properties": {
							"reason": {
								"type": "string",
								"nullable": true
							},
							"author": {
								"type": "string",
								"nullable": true
							},
							"status": {
								"type": "string",
								"enum": [
									"approved",
									"denied",
									"pending"
								]
							}
						},
						"required": [
							"reason",
							"author",
							"status"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"description",
					"authors",
					"options",
					"approval"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Files": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"AddonAll": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"authors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"options": {
						"properties": {
							"tags": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"optifine": {
								"type": "boolean"
							},
							"comments": {
								"type": "boolean"
							}
						},
						"required": [
							"tags",
							"optifine",
							"comments"
						],
						"type": "object"
					},
					"approval": {
						"properties": {
							"reason": {
								"type": "string",
								"nullable": true
							},
							"author": {
								"type": "string",
								"nullable": true
							},
							"status": {
								"type": "string",
								"enum": [
									"approved",
									"denied",
									"pending"
								]
							}
						},
						"required": [
							"reason",
							"author",
							"status"
						],
						"type": "object"
					},
					"files": {
						"$ref": "#/components/schemas/Files"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"description",
					"authors",
					"options",
					"approval",
					"files"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Textures": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Texture": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"tags"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Uses": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Paths": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Contributions": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"TextureAll": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"uses": {
						"$ref": "#/components/schemas/Uses"
					},
					"paths": {
						"$ref": "#/components/schemas/Paths"
					},
					"contributions": {
						"$ref": "#/components/schemas/Contributions"
					}
				},
				"required": [
					"id",
					"name",
					"tags",
					"uses",
					"paths",
					"contributions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"uuid": {
						"type": "string"
					},
					"muted": {
						"properties": {
							"end": {
								"type": "string"
							},
							"start": {
								"type": "string"
							}
						},
						"required": [
							"end",
							"start"
						],
						"type": "object"
					},
					"warns": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"username",
					"roles",
					"uuid",
					"muted",
					"warns"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Compliance API",
		"version": "2.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/addons/{id}": {
			"get": {
				"operationId": "GetAddon",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Addon"
								}
							}
						}
					}
				},
				"tags": [
					"Addons"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/addons/{id}/all": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddonAll"
								}
							}
						}
					}
				},
				"tags": [
					"Addons"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/addons/{id}/files": {
			"get": {
				"operationId": "GetFiles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Files"
								}
							}
						}
					}
				},
				"tags": [
					"Addons"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/textures/raw": {
			"get": {
				"operationId": "GetRaw",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Textures"
								}
							}
						}
					}
				},
				"tags": [
					"Textures"
				],
				"security": [],
				"parameters": []
			}
		},
		"/textures/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Texture"
								}
							}
						}
					}
				},
				"tags": [
					"Textures"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/textures/{id}/uses": {
			"get": {
				"operationId": "GetUses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Uses"
								}
							}
						}
					}
				},
				"tags": [
					"Textures"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/textures/{id}/paths": {
			"get": {
				"operationId": "GetPaths",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Paths"
								}
							}
						}
					}
				},
				"tags": [
					"Textures"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/textures/{id}/contributions": {
			"get": {
				"operationId": "GetContributions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Contributions"
								}
							}
						}
					}
				},
				"tags": [
					"Textures"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/textures/{id}/all": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TextureAll"
								}
							}
						}
					}
				},
				"tags": [
					"Textures"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{id}/contributions": {
			"get": {
				"operationId": "GetContributions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Contributions"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/v2"
		}
	]
}